Scratch work on figuring out the visible portion of the cell
Moves the intersectionObserver into the data call part of the code from the row part of the code.

Adds position relative to the parent data cell, so we can position the text div absolute.
