+++ src/components/tables.scss
@@ -254,3 +254,7 @@ table.draggableTableContainer {
 .portraitTableContainer {
   height: 90vh;
 }
+
+.parentData {
+  position: relative;
+}
